+++ app/Http/Controllers/LibrarianController.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class LibrarianController extends Controller
+{
+    public function dashboard(Request $request) {
+        return view('librarian.dashboard');
+    }
+}